refactor: broken up the tests for correlation-ids middleware
