+++ packages/lambda-powertools-middleware-correlation-ids/__tests__/api_gateway.js
@@ -1,0 +1,15 @@
+const _ = require('lodash')
+const { standardTests } = require('./lib')
+
+global.console.log = jest.fn()
+
+const apig = require('./event_templates/apig.json')
+const genApiGatewayEvent = (correlationIds = {}) => {
+  const event = _.cloneDeep(apig)
+  event.headers = correlationIds
+  return event
+}
+
+describe('Correlation IDs middleware (API Gateway)', () => {
+  standardTests(genApiGatewayEvent)
+})
