+++ packages/lambda-powertools-middleware-correlation-ids/__tests__/step_functions.js
@@ -1,0 +1,15 @@
+const _ = require('lodash')
+const { standardTests } = require('./lib')
+
+global.console.log = jest.fn()
+
+const sfn = require('./event_templates/sfn.json')
+const genSfnEvent = (correlationIDs = {}) => {
+  const event = _.cloneDeep(sfn)
+  event.__context__ = correlationIDs
+  return event
+}
+
+describe('Correlation IDs middleware (Step Functions)', () => {
+  standardTests(genSfnEvent)
+})
